:recycle: refactor(element): element 폴더를 관리자 메뉴에 따라 분리
