+++ src/elements/ProductManage/ProductCurrentSituation.js
@@ -1,10 +1,10 @@
-import ProductStore from '../domains/stores/ProductStore';
-import { createAction, PRODUCT_ACTION } from '../domains/actions';
+import ProductStore from '../../domains/stores/ProductStore';
+import { createAction, PRODUCT_ACTION } from '../../domains/actions';
 
-import CustomElement from '../abstracts/CustomElement';
-import { $, $$ } from '../utils/dom';
-import { checkProductValidation } from '../validators';
-import { CONFIRM_MESSAGE } from '../constants';
+import CustomElement from '../../abstracts/CustomElement';
+import { $, $$ } from '../../utils/dom';
+import { checkProductValidation } from '../../validators';
+import { CONFIRM_MESSAGE } from '../../constants';
 
 class ProductCurrentSituation extends CustomElement {
   connectedCallback() {
